requirements.txt
absl-py==1.0.0
ansiwrap==0.8.4
asynctest==0.13.0
binaryornot==0.4.4
docker-pycreds==0.4.0
docstring-parser==0.14.1
fire==0.4.0
google-api-python-client==1.12.11
google-auth==1.35.0
google-cloud-aiplatform==1.14.0
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-bigquery==2.34.2
google-cloud-bigquery-storage==2.13.0
google-cloud-bigtable==2.7.0
google-cloud-core==2.3.0
google-cloud-dataproc==4.0.1
google-cloud-datastore==2.5.1
google-cloud-firestore==2.4.0
google-cloud-kms==2.11.1
google-cloud-language==2.4.1
google-cloud-logging==3.0.0
google-cloud-monitoring==2.9.1
google-cloud-notebooks==1.3.0
google-cloud-pipeline-components==1.0.9
google-cloud-pubsub==1.7.0
google-cloud-resource-manager==1.4.1
google-cloud-scheduler==2.6.2
google-cloud-spanner==3.13.0
google-cloud-speech==2.13.1
google-cloud-storage==1.44.0
google-cloud-tasks==2.8.1
google-cloud-translate==3.7.2
google-cloud-videointelligence==2.6.1
google-cloud-vision==2.7.1
google-resumable-media==2.3.2
grpc-google-iam-v1==0.12.3
grpcio-status==1.44.0
invoke==1.7.1
jsonschema==3.2.0
keyring==23.5.0
keyrings.google-artifactregistry-auth==1.0.0
kfp==1.8.12
kfp-pipeline-spec==0.1.14
kfp-server-api==1.8.1
kubernetes==18.20.0
missingno==0.4.2
munkres==1.1.4
pandas==1.3.5
pluggy==1.0.0
poyo==0.5.0
proto-plus==1.20.3
protobuf==3.19.4
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
PyYAML==5.4.1
requests-toolbelt==0.9.1
retrying==1.3.3
strip-hints==0.1.10
tabulate==0.8.9
termcolor==1.1.0
text-unidecode==1.3
textwrap3==0.9.2
typer==0.4.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
webencodings==0.5.1
